+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* [minimagick] Automatically sharpen thumbnails after resizing (@janko-m, @mokolabs)
+
 * [vips] Automatically sharpen thumbnails after resizing (@janko-m, @mokolabs)
 
 ## 0.11.2 (2018-03-31)