+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## HEAD
+## 1.0.0 (2018-04-04)
 
 * Depend on `mini_magick` and `ruby-vips` gems (@janko-m, @mokolabs)